made touch ups for review
