+++ client/src/lib/factorDefaults.js
@@ -1,0 +1,10 @@
+const factorDefaults = {
+  Size: { value: 0, text: 'none selected' },
+  Width: { value: 0, text: 'none selected' },
+  Comfort: { value: 0, text: 'none selected' },
+  Quality: { value: 0, text: 'none selected' },
+  Length: { value: 0, text: 'none selected' },
+  Fit: { value: 0, text: 'none selected' },
+};
+
+export default factorDefaults;
