+++ public/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <title>Project Atelier</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <!-- <link rel="icon" href="favicon.png"> -->
     <!-- <link href="https://fonts.googleapis.com/css2?family=M+PLUS+Rounded+1c&display=swap" rel="stylesheet"> -->
     <script defer src="bundle.js"></script>
   </head>
